Add flag to define custom method body
There are cases in which would be useful to be able to define a custom method body.
For example, one would like to be able to print to a custom logger or generate comments with `TODO` instructions.

This commit introduces a new flag called `-body` which will receive a path to a file containing the method body definition.