Fix panic.
Had a lingering type assertion that should have used the ID from the
parsed AST type instead.

Also noticed that we were losing the package names on the type
parameters, which would lead to potentially-subtle incorrect results; we
should always be using the package if the user submits it. This requires
an update to typeFromAST to preserve the package, and then updating
findInterface to strip the package off when dealing with qualified
interfaces.